added simple testbench for vector and improve the ignore file

+++ UniPrint/print_implem.h
@@ -1,13 +1,10 @@
-
 #pragma once
-
-#include <typeinfo.h>
 
 template <typename T>
 void
 UniPrint::p(T &s)
 {
-	printf("%s[%d]*%d:\n", typeid(s).name(), &s, s.size());
+	printf("%s[%d]*%d:\n", typeid(s).name(), (int)&s, s.size());
 	s.traverse(print);
 	printf("\n");
 }
